+++ src/main/java/lotto/controller/LottoController.java
@@ -1,4 +1,11 @@
-package lotto;
+package lotto.controller;
+
+import lotto.domain.LottoTickets;
+import lotto.domain.MoneyManager;
+import lotto.domain.Ranks;
+import lotto.domain.WinningNumbers;
+import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class LottoController {
