+++ src/test/java/lottoTest/domain/LottoTicketTest.java
@@ -1,13 +1,13 @@
-package lottoTest;
+package lottoTest.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.LottoTicket;
-import lotto.Rank;
-import lotto.WinningNumbers;
+import lotto.domain.LottoTicket;
+import lotto.domain.Rank;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
