+++ src/test/java/lottoTest/domain/RanksTest.java
@@ -1,11 +1,11 @@
-package lottoTest;
+package lottoTest.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.EnumMap;
 import java.util.List;
-import lotto.Rank;
-import lotto.Ranks;
+import lotto.domain.Rank;
+import lotto.domain.Ranks;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
