+++ src/test/java/lottoTest/view/InputViewTest.java
@@ -1,10 +1,10 @@
-package lottoTest;
+package lottoTest.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.InputView;
+import lotto.view.InputView;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
